LayoutTests/svg/animations/svgenum-animation-7-expected.txt

SVG 1.1 dynamic animation tests

Test SVGStitchOptions/TurbulenceType enumeration animations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS turbulence.type.animVal is SVGFETurbulenceElement.SVG_TURBULENCE_TYPE_FRACTALNOISE
PASS turbulence.type.baseVal is SVGFETurbulenceElement.SVG_TURBULENCE_TYPE_FRACTALNOISE
PASS turbulence.stitchTiles.animVal is SVGFETurbulenceElement.SVG_STITCHTYPE_STITCH
PASS turbulence.stitchTiles.baseVal is SVGFETurbulenceElement.SVG_STITCHTYPE_STITCH
PASS turbulence.type.animVal is SVGFETurbulenceElement.SVG_TURBULENCE_TYPE_FRACTALNOISE
PASS turbulence.type.baseVal is SVGFETurbulenceElement.SVG_TURBULENCE_TYPE_FRACTALNOISE
PASS turbulence.stitchTiles.animVal is SVGFETurbulenceElement.SVG_STITCHTYPE_STITCH
PASS turbulence.stitchTiles.baseVal is SVGFETurbulenceElement.SVG_STITCHTYPE_STITCH
PASS turbulence.type.animVal is SVGFETurbulenceElement.SVG_TURBULENCE_TYPE_TURBULENCE
PASS turbulence.type.baseVal is SVGFETurbulenceElement.SVG_TURBULENCE_TYPE_FRACTALNOISE
PASS turbulence.stitchTiles.animVal is SVGFETurbulenceElement.SVG_STITCHTYPE_NOSTITCH
PASS turbulence.stitchTiles.baseVal is SVGFETurbulenceElement.SVG_STITCHTYPE_STITCH
PASS turbulence.type.animVal is SVGFETurbulenceElement.SVG_TURBULENCE_TYPE_TURBULENCE
PASS turbulence.type.baseVal is SVGFETurbulenceElement.SVG_TURBULENCE_TYPE_FRACTALNOISE
PASS turbulence.stitchTiles.animVal is SVGFETurbulenceElement.SVG_STITCHTYPE_NOSTITCH
PASS turbulence.stitchTiles.baseVal is SVGFETurbulenceElement.SVG_STITCHTYPE_STITCH
PASS turbulence.type.animVal is SVGFETurbulenceElement.SVG_TURBULENCE_TYPE_FRACTALNOISE
PASS turbulence.type.baseVal is SVGFETurbulenceElement.SVG_TURBULENCE_TYPE_FRACTALNOISE
PASS turbulence.stitchTiles.animVal is SVGFETurbulenceElement.SVG_STITCHTYPE_STITCH
PASS turbulence.stitchTiles.baseVal is SVGFETurbulenceElement.SVG_STITCHTYPE_STITCH
PASS successfullyParsed is true

TEST COMPLETE
